+++ demo/app/home/utilities.ts
@@ -1,0 +1,11 @@
+import { android } from "tns-core-modules/application";
+
+export const getDeepLink = (path = "") => {
+  const scheme = 'my-demo';
+  const prefix = android ? `${scheme}://demo/` : `${scheme}://`;
+  return prefix + path;
+};
+
+export const sleep = (timeout: number) => {
+  return new Promise(resolve => setTimeout(resolve, timeout));
+};