+++ src/index.d.ts
@@ -35,7 +35,8 @@ declare module 'nativescript-inappbrowser' {
       | 'crossDissolve'
       | 'partialCurl',
     modalEnabled?: boolean,
-    enableBarCollapsing?: boolean
+    enableBarCollapsing?: boolean,
+    ephemeralWebSession?: boolean
   };
 
   type InAppBrowserAndroidOptions = {